+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+- Fixed namespacing bug when an Address field is nested within a Matrix field. ([#43](https://github.com/doublesecretagency/craft-googlemaps/issues/43))
+
 ## 4.1.0 - 2021-12-01
 
 > {warning} If you are using [marker clustering](https://plugins.doublesecretagency.com/google-maps/dynamic-maps/clustering-markers/), please be aware that the internal library has changed and your code may require minor adjustments as a result.